[#53][v2.1.2] Download RSS feed from plugin core
Currently, RSS feeds are fetched from the plugin core on the server
when running a subscription automatically by the scheduler, and
directly from the GTK client when previewing in the subscription
dialog. This can cause confusion if the previewing works, but not
throught the scheduler.

Fix by always fetching RSS feeds via the plugin core